fix(seo): split sitemap into <50k chunks via generateSitemaps
O /sitemap.xml monolitico gerava ~66k URLs, acima do limite de 50k do
Google, que rejeitava o arquivo inteiro. Agora:

- Logica de URLs extraida para lib/sitemap-entries.ts (fonte unica)
- app/sitemap.ts usa generateSitemaps() -> /sitemap/{id}.xml (chunks 40k)
- sitemap-index.xml enumera os chunks + sub-sitemaps paginados /api/*
- robots.ts remove referencia morta a /sitemap.xml (agora e o index)
- public/robots.txt (legado) alinhado: aponta ao index, libera /api/sitemap/

Typecheck do web nao introduz nenhum erro novo.
